[Decomp][Player] Cohesion split: extract visibility methods to PlayerVisibility.cpp
